app/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-admin-sortable2==0.7.8
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-tag-parser==3.2
djangorestframework==3.12.2
future==0.18.2
kombu==5.0.2
prompt-toolkit==3.0.17
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
